// shipeasy/src/app/models/party-master.ts
export interface PartyMasterData {
    partyShortcode:any;
    tax_number: string;
    _id: {
      $oid: string;
    };
    partymasterId: string;
    ImportExport: string;
    TDSNature: string;
    TDS_limit: string;
    TDS_utilised: string;
    accountCode: string;
    activeflag: boolean;
    adCode: string;
    addressInfo: {
      address: string;
      countryId: string;
      countryISOCode: string;
      countryName: string;
      stateId: string;
      stateName: string;
      cityId: string;
      cityName: string;
      postalCode: string;
      stateCode: string;
    };
    annualTurnover: string;
    bankName: string;
    branch: {
      branch_name: string;
      branch_address: string;
      barnch_country: string;
      branch_countryName: string;
      branch_state: string;
      branch_stateName: string;
      branch_cityId: string;
      branch_city: string;
      pinCode: string;
      stateCodeBranch: string;
      placeofSupply: string;
      partyCode: string;
      bankName: string;
      cust_acc_no: number;
      remarks: string;
      pic: string;
      pic_phone: number;
      pic_email: string;
      tax_name: string;
      tax_number: string;
      BROKERAGE_PAYABLE: boolean;
      creadirCustomer: boolean;
      active_flag: boolean;
      kyc_flag: boolean;
      TDS_GST_APPLICABLE: boolean;
      documents: any; // Replace 'any' with appropriate type if known
    }[];
    brokerageAccCode: string;
    brokeragePayable: boolean;
    chaId: string;
    checkAcceptStatus: boolean;
    chequeAcceptance: string;
    createdBy: string;
    createdOn: string;
    creditCustomer: boolean;
    creditDays: string;
    currency: any;
    partyCurrency:any // Replace 'any' with appropriate type if known
    customerAccCode: string;
    customerAccNumber: string;
    customerName: string;
    customerStatus: string;
    customerType: {
      item_id: string;
      item_text: string;
    }[];
    deactivationDate: string;
    deactivationReason: string;
    discountedBlFees: number;
    expFlag: boolean;
    expScCode: string;
    exportTeun: string;
    fasCode: string;
    faxNo: {
      faxCountryCode: string;
      faxAreaCode: string;
      faxNo: string;
    };
    flatRateDetention: string;
    groupCompany: boolean;
    gstNo: string;
    impFlag: boolean;
    impScCode: string;
    importTeun: string;
    isSez: string;
    isShippingLine: boolean;
    isTDSExempt: boolean;
    kycFlag: boolean;
    location: string;
    locationId: string;
    mloId: string;
    modifiedBy: string;
    module: string;
    name: string;
    operationUser: string;
    opsCode: string;
    opsFrom: string;
    opsLocation: string;
    opsName: string;
    orgId: string;
    panNo: string;
    primaryMailId: string;
    primaryNo: {
      primaryCountryCode: string;
      primaryAreaCode: string;
      primaryNo: string;
    };
    principle: string;
    proformaRequiredExp: boolean;
    proformaRequiredImp: boolean;
    rateCheckBL: boolean;
    rateCheckBooking: boolean;
    rateCheckDraftBL: boolean;
    rateCheckRO: boolean;
    refundAccCode: string;
    reg_UN_UIN: string;
    remarks: string;
    residentStatus: boolean;
    sacCode: string;
    saleCode: string;
    saleFrom: string;
    saleLocation: string;
    saleName: string;
    salePrinciple: string;
    serviceTax: string;
    serviceTaxExmp: string;
    serviceTaxExmpNo: number;
    serviceTaxLastDate: string;
    serviceTaxLastModify: string;
    serviceTaxNo: number;
    shortName: string;
    status: boolean;
    tanNo: string;
    tdsCertificate: boolean;
    tdsCertificateValid: string;
    tdsFlag: boolean;
    tdsGSTApplicable: boolean;
    tdsPercent: number;
    tenantId: string;
    tosite: string;
    updatedBy: string;
    updatedOn: string;
    url: string;
    amount: number;
    customer: any[]; // Replace 'any' with appropriate type if known
    lumsumDateFrom: string;
    lumsumDateTo: string;
    isRegister: boolean;
    daysinPeriod: number;
    updatedByUID: string;
  }